refactor(object): extract descriptor introspection to descriptors.rs (keep object_ops.rs under 2000-line cap)
